Kafka-Cluster-Topic-producer-consumer created
Project Description: Kafka Cluster Topic Producer, REST API, and Consumer

This project showcases a Kafka cluster setup with a topic producer, REST API, and consumer implemented in Spring Boot. It serves as a foundation for developing real-time event streaming and messaging applications using Kafka. The project demonstrates the basic setup and usage of Kafka's producer and consumer APIs.

To run the applications:

1. Start Kafka Locally: Open separate terminals and run the following commands:

* Start ZooKeeper: .\bin\windows\zookeeper-server-start.bat .\config\zookeeper.properties
* Start Kafka Server: .\bin\windows\kafka-server-start.bat .\config\server.properties
*  Start Message Consumer: bin\windows\kafka-console-consumer.bat --bootstrap-server localhost:9093 --topic javaguides --from-beginning

2. Run the Application: Launch the application from your preferred IDE (e.g., Eclipse, IntelliJ, etc.).

3. Access the API: Open a web browser and visit http://localhost:8080/api/v1/kafka/publish?message=add a message here. 
This allows you to publish a message via the REST API.

Through this setup, you can observe message consumption and production using the provided Kafka consumers and producers.